feat(ui): DOMA-8189 rename Tip to TourStep, changed inner tooltip to popover, fix arrow indents
