--- packages/ui/src/components/Tip/index.ts
@@ -1,5 +1,0 @@
-import { Tip } from './tip'
-import './style.less'
-
-export type { TipProps } from './tip'
-export { Tip }
