+++ packages/ui/src/components/TourStep/index.ts
@@ -1,0 +1,5 @@
+import { TourStep } from './tourStep'
+import './style.less'
+
+export type { TourStepProps } from './tourStep'
+export { TourStep }
